requirements11.txt
flask
python-dotenv
supabase
flask-cors